Handle Angular routes in Spring
When navigating directly to an Angular route, Spring won't return index.html by default.  To work around this, we use a custom resource resolver that returns the content of index.html for any resource path that doesn't end in a file extension (i.e. that doesn't look like it's a reference to a real static resource).